+++ app/src/main/java/com/fkexample/cointicker/ui/composables/CyptoImage.kt
@@ -1,4 +1,4 @@
-package com.fkexample.cointicker.ui
+package com.fkexample.cointicker.ui.composables
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
