+++ app/src/main/java/com/fkexample/cointicker/ui/composables/LoadingCryptoListShimmer.kt
@@ -1,4 +1,4 @@
-package com.fkexample.cointicker.ui
+package com.fkexample.cointicker.ui.composables
 
 import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.RepeatMode
